fix: create a card component for all the cards, fix username overflow on mobile card, fix hottest page padding
