+++ src/features/Profiles/pages/EditProfilePage/AccountCard/AccountCard.tsx
@@ -1,6 +1,7 @@
 import Button from 'src/Components/Button/Button';
 import { useAppDispatch } from 'src/utils/hooks';
 import { openModal } from 'src/redux/features/modals.slice';
+import Card from 'src/Components/Card/Card';
 
 
 interface Props {
@@ -17,10 +18,8 @@ export default function AccountCard({ }: Props) {
 
 
     return (
-        <div className="rounded-16 bg-white border-2 border-gray-200 p-24">
-            <p className="text-body2 font-bold">Account Settings</p>
-
-
+        <Card>
+            <p className="text-body2 font-bold">🔒 Linking Accounts</p>
             <div className='mt-24 flex flex-col gap-16'>
                 <p className="text-body3 font-bold">Linked Wallets</p>
                 <p className="text-body4 text-gray-600">
@@ -32,8 +31,6 @@ export default function AccountCard({ }: Props) {
                     Connect new wallet ⚡
                 </Button>
             </div>
-
-
-        </div>
+        </Card>
     )
 }
